# Fast pipeline binding for ForEach-Object