@PrzemyslawKlys 2019-07-15
I have an open issue https://github.com/PowerShell/PowerShell/issues/9941 for Where-Object. @daxian-dbw is taking care of it. Maybe those two can be worked together? Seems similar to me?