@iSazonov 2019-07-15
I guess there is 2 places for perf consideration:
- Bindings. If source collection is strong typed collection we could bind only once. For common case source collection may contains free-typed objects and we have to do binding for every object.
- Compilation. Perhaps we could compile to something more faster. Ex., for Where-Object a filter function is more faster than scriptblock, maybe a delegate is better. Although maybe this is again binding.