@mi-hol 2020-03-22
@lzybkr @daxian-dbw could the suggestion from @TobiasPSP  be considered for 7.1? In his blog he shows many alternative approaches to speed up pwsh that should be unnecessary due to optimizations in the core.