@daxian-dbw 2020-03-26
@mi-hol Please check out the issue https://github.com/PowerShell/PowerShell/issues/10982. There is tons of discussion there.